refactor: domain 패키지를 tobe 아래로 이동
