+++ src/main/java/kitchenpos/products/tobe/domain/JpaProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.tobe.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
