+++ src/test/java/kitchenpos/products/tobe/domain/ProductTest.java
@@ -1,7 +1,6 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.tobe.domain;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
